% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification_tests.R
\name{bootstrap_specification_tests}
\alias{bootstrap_specification_tests}
\title{Call metastudy_X_sigma_cors repeatedly with new bootstrap samples to
compute standard errors for correlations using the inverse probability weighting approach}
\usage{
bootstrap_specification_tests(X, sigma, B = 10, ..., num.cores = 1)
}
\arguments{
\item{X}{vector of parameters}

\item{sigma}{vector of standard errors}

\item{B}{number of bootstrap replications}

\item{...}{parameters passed to \code{metastudies_estimation}}

\item{num.cores}{number of cores used for parallel computation using \code{mclapply}.}
}
\description{
Note that this procedure can take very long to run.
}
