% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastudiesplots.r
\name{estimates_plot}
\alias{estimates_plot}
\title{Plot}
\usage{
estimates_plot(
  ms,
  X = ms$X,
  sigma = ms$sigma,
  cutoffs = ms$cutoffs,
  symmetric = ms$symmetric,
  model = ms$model
)
}
\description{
Plot
}
