% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastudiesfunctions.r
\name{metastudies_estimation}
\alias{metastudies_estimation}
\title{Perform Andrews and Kasy (2019) estimation}
\usage{
metastudies_estimation(
  X,
  sigma,
  cutoffs = c(1.96),
  symmetric = FALSE,
  model = "normal",
  eval.max = 10^5,
  iter.max = 10^5,
  abs.tol = 10^(-8),
  stepsize = 10^(-6)
)
}
\arguments{
\item{X}{vector of reported coefficients}

\item{sigma}{vector of reported stanard errors}

\item{cutoffs}{significance thresholds that define intervals of different publication probabilities, e.g. \code{c(1.645, 1.96, 2.576)}.}

\item{symmetric}{if TRUE assume that positive and negative z-statistics are symmetrically distributed.}

\item{model}{either "normal" or "t". The assumed functional form of the distribution absent publication bias.}

\item{eval.max, iter.max, abs.tol}{Control parameters for [stats::nlminb].}
}
\description{
Perform Andrews and Kasy (2019) estimation
}
