% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{MetaStudiesApp}
\alias{MetaStudiesApp}
\title{Create and run the MetaStudies shiny app}
\usage{
MetaStudiesApp(csv.file = NULL, show.cor = TRUE, ...)
}
\arguments{
\item{...}{additional parameters passed to \code{shinyApp}}
}
\description{
Create and run the MetaStudies shiny app
}
